requirements.txt
customtkinter
speechrecognition
pyttsx3
pywhatkit
wikipedia
pyjokes
pyaudio
openai
requests
pillow
playsound
